rename include folder for consistency with 'gtest' and 'gmock'
