+++ include/gtest-verilog/ConsoleColour.h
