+++ include/gtest-verilog/MatchesTrace.h
