+++ include/gtest-verilog/Trace.h
