+++ include/gtest-verilog/TraceBuilder.h
