+++ test/TraceBuilder.test.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 using namespace testing;
 
-#include "googletest-verilog/googletest-verilog.h"
+#include "gtest-verilog/gtest-verilog.h"
 using namespace gtestverilog;
 
 namespace {
